🐛(autoconfig): Fix ConditionalOnClass to annotate type
