+++ .idea/runConfigurations/All_in_selidor_autoconfigure.xml
@@ -22,6 +22,7 @@
     <option name="MAIN_CLASS_NAME" value="" />
     <option name="METHOD_NAME" value="" />
     <option name="TEST_OBJECT" value="package" />
+    <option name="VM_PARAMETERS" value="-ea --add-modules org.slf4j --add-opens org.slf4j/org.slf4j=ALL-UNNAMED" />
     <option name="PARAMETERS" value="" />
     <option name="WORKING_DIRECTORY" value="$PROJECT_DIR$" />
     <method v="2">
