+++ selidor-projects/selidor-autoconfigure/src/main/java/pw/itr0/selidor/boot/autoconfigure/http/client/proxy/HttpProxyAutoConfiguration.java
@@ -3,7 +3,6 @@ package pw.itr0.selidor.boot.autoconfigure.http.client.proxy;
 import java.net.ProxySelector;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.function.Consumer;
 import org.springframework.boot.autoconfigure.AutoConfigureAfter;
 import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
@@ -11,14 +10,10 @@ import org.springframework.boot.autoconfigure.condition.ConditionalOnClass;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.context.PropertyPlaceholderAutoConfiguration;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.boot.web.reactive.function.client.WebClientCustomizer;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
 import org.springframework.core.Ordered;
-import org.springframework.http.client.reactive.ClientHttpConnector;
-import org.springframework.http.client.reactive.ReactorClientHttpConnector;
-import org.springframework.http.client.reactive.ReactorResourceFactory;
-import org.springframework.web.reactive.function.client.WebClient;
 import pw.itr0.selidor.boot.autoconfigure.http.client.proxy.HttpProxyConfigurationProperties.ProxySetting;
 import pw.itr0.selidor.boot.autoconfigure.http.client.proxy.HttpProxyConfigurationProperties.ProxySetting.BasicAuthenticationSetting;
 import pw.itr0.selidor.http.client.Scheme;
@@ -26,16 +21,13 @@ import pw.itr0.selidor.http.client.proxy.EnvVarProxyLoader;
 import pw.itr0.selidor.http.client.proxy.HttpProxies;
 import pw.itr0.selidor.http.client.proxy.HttpProxy;
 import pw.itr0.selidor.http.client.proxy.HttpProxy.Builder;
-import reactor.netty.http.client.HttpClient;
-import reactor.netty.tcp.ProxyProvider;
-import reactor.netty.tcp.ProxyProvider.Proxy;
-import reactor.netty.tcp.ProxyProvider.TypeSpec;
 
 @Configuration(proxyBeanMethods = false)
 @AutoConfigureOrder(Ordered.HIGHEST_PRECEDENCE)
 @AutoConfigureAfter(PropertyPlaceholderAutoConfiguration.class)
 @ConditionalOnClass(HttpProxies.class)
 @EnableConfigurationProperties(HttpProxyConfigurationProperties.class)
+@Import(HttpProxyConfiguration.ReactorWebClientProxyConfiguration.class)
 public class HttpProxyAutoConfiguration {
   private static final List<ProxyEnvVar> PROXY_ENV_VARS =
       List.of(
@@ -74,41 +66,6 @@ public class HttpProxyAutoConfiguration {
   private int getProxiesSize() {
     return properties.getProxies().size()
         + (properties.isUseEnvironmentVariables() ? PROXY_ENV_VARS.size() : 0);
-  }
-
-  @Bean
-  @ConditionalOnClass({WebClient.class, HttpClient.class})
-  @ConditionalOnBean({HttpProxies.class, ReactorResourceFactory.class})
-  @ConditionalOnMissingBean(ClientHttpConnector.class)
-  WebClientCustomizer webClientProxyCustomizer(
-      HttpProxies proxies,
-      @SuppressWarnings("SpringJavaInjectionPointsAutowiringInspection") // たぶん、IntelliJのバグじゃないかと。
-          ReactorResourceFactory factory) {
-    return proxies.primary().map(p -> applyProxy(factory, p)).orElse(b -> {});
-  }
-
-  private WebClientCustomizer applyProxy(ReactorResourceFactory factory, HttpProxy p) {
-    return builder ->
-        builder.clientConnector(
-            new ReactorClientHttpConnector(
-                factory,
-                client ->
-                    client.tcpConfiguration(
-                        tcpClient -> tcpClient.proxy(webClientProxyBuilder(p)))));
-  }
-
-  private Consumer<TypeSpec> webClientProxyBuilder(HttpProxy p) {
-    return typeSpec -> {
-      final ProxyProvider.Builder spec =
-          typeSpec
-              .type(Proxy.HTTP)
-              .host(p.getHost())
-              .port(p.getPort())
-              .nonProxyHosts(p.getNonProxyHostsRegex());
-      if (p.isAuthenticationRequired()) {
-        spec.httpHeaders(entries -> p.getAuthorizationHeaders().forEach(entries::add));
-      }
-    };
   }
 
   private HttpProxy createNetworkProxy(ProxySetting proxySetting) {
