+++ selidor-projects/selidor-autoconfigure/src/main/java/pw/itr0/selidor/boot/autoconfigure/jackson/SelidorJacksonAutoConfiguration.java
@@ -10,12 +10,11 @@ import org.springframework.core.annotation.Order;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
 
 @Configuration(proxyBeanMethods = false)
-@ConditionalOnClass(ObjectMapper.class)
+@ConditionalOnClass({ObjectMapper.class, Jackson2ObjectMapperBuilder.class})
 public class SelidorJacksonAutoConfiguration {
   @Bean
   // Spring BootのApplication PropertiesでのJacksonの設定が有効になるように、Listの先頭に入れる。
   @Order(Ordered.HIGHEST_PRECEDENCE)
-  @ConditionalOnClass(Jackson2ObjectMapperBuilder.class)
   Jackson2ObjectMapperBuilderCustomizer selidorStandardJacksonConfiguration() {
     return new SelidorStandardJacksonConfiguration();
   }
